blimp-safety-env/requirements.txt
absl-py==2.0.0
# # Editable Git install with no remote (blimp-env==0.1.0)
# -e /Users/dbarker9/Projects/blimp_simple/blimp-env
cachetools==5.3.2
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
contourpy==1.2.0
cycler==0.12.1
Farama-Notifications==0.0.4
filelock==3.13.1
fonttools==4.44.0
fsspec==2023.10.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
grpcio==1.59.2
gymnasium==0.29.1
idna==3.4
importlib-resources==6.1.1
Jinja2==3.1.2
kiwisolver==1.4.5
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.1
mdurl==0.1.2
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.2
oauthlib==3.2.2
opencv-python==4.8.1.78
packaging==23.2
Panda3D
pandas==2.1.3
Pillow==10.1.0
protobuf==4.23.4
psutil==5.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pygame==2.5.2
Pygments==2.16.1
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.6.0
rsa==4.9
sb3-contrib==2.1.0
scipy==1.11.3
Shimmy==1.1.0
six==1.16.0
stable-baselines3==2.1.0
sympy==1.12
tabulate==0.9.0
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
Werkzeug==3.0.1
moviepy==1.0.3
loguru==0.7.2
pyyaml==6.0.1
pyarrow==15.0.0